Fix VC9 Compiler warning C4267: conversion from 'size_t' to 'const int', possible loss of data
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@8564 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/FragmentedFile.cpp
@@ -303,8 +303,8 @@ BOOL CFragmentedFile::Open(const CBTInfo& oInfo, const BOOL bWrite,
 	CString& strErrorMessage)
 {
 	CString sUniqueName = oInfo.GetFilename();
-	int i = 0;
-	const int nCount = m_oFile.size();
+	size_t i = 0;
+	const size_t nCount = m_oFile.size();
 	QWORD nOffset = 0;
 
 	for ( POSITION pos = oInfo.m_pFiles.GetHeadPosition() ; pos ; ++i )
